// src/platforms/session/enums.ts
export enum SessionState {
  INITIAL = "INITIAL",
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
  LOGGED_OUT = "LOGGED_OUT",
  REFRESH = "REFRESH",
  CHECKING = "CHECKING",
}

export enum WorkerEventType {
  START_INACTIVITY = "START_INACTIVITY",
  RESET_INACTIVITY = "RESET_INACTIVITY",
  START_REFRESH = "START_REFRESH",
  STOP_REFRESH = "STOP_REFRESH",
  DESTROY = "DESTROY",
  INACTIVE = "INACTIVE",
  REFRESH = "REFRESH",
}